util: Updated tsan_process_trace to handle "thread start/end" events
The tool now handles thread start/end events and uses that to assign thread IDs
that TSan will see. Reuse of the "real thread IDs" (addresses of task_struct in
the kernel) is now handled properly.

tsan_process_trace now also outputs the list of threads, with names (based on
task_struct->comm[]), at the end of the report. This may make the report easier
to understand.