util: Fixed: init functions of some target modules were not determined properly
Before this commit, tsan_process_trace tool assumed that the first function in
the init area of a target module is its init function. When the init function
ends, the tool marked the init area as free and thus handled its reuse by other
modules.

There are modules, however, which init functions are not the first in the init
areas (e.g. fuse in kernel 3.4.28-desktop on OpenSUSE 12.2 x86). They confused
tsan_process_trace when it saw the events from the real init function and could
not find a module they belong to.

This is now fixed. A test has also been added to detect regressions of this bug
in the future.