func: Added a mechanism to pass data to the callback handlers
It turned out to be convenient to save the pointer to the data that can be
used by pre- and post- handlers for the functions in kedr_func_info along
with the handlers themselves.

kedr_func_info::data can now be used for this purpose. The rules for
accessing it are the same as for the handlers: RCU-aware, setting under
handler_lock, reading with rcu_dereference().

As the execution of the handlers takes place under rcu_read_lock(), the
handlers do not need to use rcu_read_lock() / rcu_read_unlock() themselves
when accessing the data but rcu_dereference() is needed. Or at least,
appropriate memory barriers should be used if the pointer is used but the
data pointed to are not.