tests: Removed restriction on thread ID in the reporter tool
Before this commit, if the reporter tool was given a function to process,
only the events from the thread executing that function were reported. This
turned out to be inconvenient when testing how delayed execution stuff
works (timers, workqueues, etc.) The callbacks involved there are usually
invoked in a thread different from the one that has set them.

Now all the events that occurred since a thread entered the function till it
has exited it will be recorded no matter which thread they belong to.