core: Made the IDs returned by kedr_get_unique_id() last "forever"
The IDs of happens-before arcs returned by kedr_get_unique_id() are now valid
until KernelStrider is unloaded rather than until the session with the targets
ends.

This makes it easier to use such IDs for the happens-before relations involving
the objects that belong to the kernel proper rather than to a target module
(e.g. system workqueue objects).