func: Revisited the testing facilities for the handlers of kernel functions
To make it more convenient to write the tests for the handlers of kernel
functions, the testing facilities have been updated.

1. The user may now specify a shell script to check the trace rather than AWK
script (see TEST_CHECKER_SCRIPT in CMakeLists.txt). This script gets the name
of the target function as the first argument and the path to the trace file as
the second one. This allows to check different things in the trace for the
different functions.

2. A checker script has access to CMAKE_SOURCE_DIR and CMAKE_BINARY_DIR as
environment variables. This allows to provide such scripts "as is" most of the
time, without the need to configure them.

The tests checking the handlers for the timer API have been updated to make use
of this new model.