util: Implemented debug mode in tsan_process_trace tool
Debug mode is turned on if "--debug" is specified as an argument to the
tool. It can be used to debug the software that has collected the trace as
well as tsan_process_trace itself.

In this mode, the tool does not invoke ThreadSanitizer but rather outputs
the trace that would have been passed to it to stdout. Note that the
addresses of the code locations in the trace are effective addresses.
tsan_process_trace chooses the addresses for "init" and "core" memory areas
of each target module and recalculates other addresses based on these. This
allows to ensure that each time a given target module is loaded during an
analysis session, the addresses of the locations in its code will be the
same in the trace no matter where in the memory it is actually located.

The addresses of the data are always writted to the trace as they are,
without any changes.

Additional information (effective addresses of the sections in the target
modules, etc.) will be output to stderr.