func: Reworked the build system
The build system for FH plugins has been redesigned and generalized. Now, it
can be used not only for the plugin for common operations.

The most significant changes are as follows.

1.
The groups of functions may specify the source files they provide for the
plugin. Before this commit, each group could provide only handlers_*.c. This
was not sufficient if, for example, not only the handlers for the exported
kernel functions but also the handlers for the callbacks are generated using
the templates (needed for netdev ops, PCI ops, etc.).

Now each group may specify any source files it provides as is or generates
and they will be used when building the plugin module.

2.
Group-specific details (definitions of the handler structures, etc.) are now
encapsulated within the groups' source files and no longer litter the main
source file of the plugin.

Among other things, the revisited build system allowed that main source file
to remain static (rather than generated), which simplifies debugging. The
code is also cleaner now.

The main source file only needs to list the groups of functions it handles
appropriately (see the description of KEDR_FH_DECLARE_GROUP() and
KEDR_FH_ADD_GROUP() in functions/util/fh_plugin.h and the code of
functions/common/module.c)

3.
The changes also allowed to get rid of header.data files in the function
groups as well as of some other stuff.