func: Simplified building the FH plugin from different source files
The previously used group_set_sources() was too clumsy to use and did not
help much if the sources were in deeper subdirectories (add_custom_target()
only sees add_custom_command() in the same directory where the former is
defined).

Now it is enough to call group_add_sources() in each subdirectory that
provides the sources to the FH plugin and list the sources there (names
only, without the path). Necessary targets and dependencies will be created
automatically.