func: Automated handling of function groups a bit more
Before this commit, one needed to manually add special macros for a group
in module.c when a group was added. This was too easy to forget.

module.c is now generated automatically from the templates, so no manual
actions are needed w.r.t. module.c when adding a new function group.