func: Added API for lock accounting
Handling of the callbacks that execute under a lock taken by the kernel or by a
target module itself requires keeping track of which locks does the target
module lock.

This is needed to avoid nested lock/unlock events for the same lock in the
callbacks (a callback might call another one explicitly as well, etc.).