util: Provided a helper script (kedr.py) to load/unload kernel-mode components
It is now possible, for example, to load all kernel-mode components of
KernelStrider in an appropriate order in a single step:

  /usr/local/bin/kedr.py start --tools=KernelStrider --targets=<list_of_targets>

If 'kedr.py start' was used to start the tools, 'kedr.py stop' should be used to
stop them. The list of modules is stored in '/run/kedr-running' by the former
command and read from there by the latter.

See '/usr/local/bin/kedr.py --help' for details.

Requires Python 2.7.