func: pci: Express that pci_driver callbacks are synchronized
pci_driver callbacks (.probe, .remove, ...) are executed under device_lock()
mutex, so are dev_pm_ops. This is now expressed approximately with appropriate
happens-before relations.

pci_driver::sriov_configure() seems to be an exception but it is rarely used at
the moment. This difference in synchronization rules is just ignored here for
simplicity.