net: Added handling of skb lifetime (simplified)
A number of situations when socket buffer structures (skb) are allocated and
frees are now tracked.

To be exact, only allocation/deallocation of struct sk_buff itself is tracked
but not the data its members refer to (the frame data itself). This is done for
simplicity and to make maintenance easier.

Tracking skb lifetime may help reduce the number of false positives that arise
due to memory reuse.