core: Fixed handling of recursive calls
Recursive function calls are rare in the kernel modules and are often frowned
upon. Still they can be and are, in fact, used. In 'vboxdrv' module from
VirtualBox 4.3.2, for example, rtR0MemObjNativeGetPagePhysAddr() calls itself
with different arguments in some cases.

Before this fix, such calls were treated as the local control transfers rather
than the calls to other functions. This lead to problems: the local storage was
clobbered by the call, therefore, incorrect register values were used and system
 crashed in a hard-to-diagnose way as a result.

BTW, this is the root cause of the problem addressed by the previous commit
(d471b435957b).

Should be fixed now. I also added a test for handling of recursive calls.