core: Revisited annotation-related parts
The goal is to be able to add the support for KEDR annotations to the kernel
itself if needed. This seems to be more useful when analyzing in-tree
modules, especially the groups of such modules, rather than embedding the
annotation functions into a particular target module.

A patch or a series of patched to the kernel are to be created.

To facilitate such use cases, the following has been done already.

1. The preprocessor symbol to turn on/off the annotations is
CONFIG_KEDR_ANNOTATIONS now. If it is set to a non-zero value, the
annotations (KEDR_ANNOTATE_*()) will be available, otherwise they will be
no-ops. No code is added to the kernel/driver when the annotations are not
enabled.

As a consequence, it is no longer possible to enable annotations for some
source files of a kernel module but do not enable them for other source
files. This was not very useful, anyway.

2. The test for the annotations is now built differently depending on
whether the annotations are enabled in the kernel. If they are, the test
module will use these, otherwise the necessary functions will be built into
the test module as before.