core, func: In-kernel annotations should now be handled
The KEDR annotations built into the target module were already handled before.

This commit adds handling of such annotations provided by the kernel itself.
They are treated in a similar way as other exported kernel functions.

Among other things, handling of the function groups that could be empty was
fixed in the process.