Added miniLZO to the project
miniLZO is a subset of the library that provides API for LZO compression and
decompression. It will be used by the user-space trace processing tools.
Taken from http://www.oberhumer.com/opensource/lzo/.

This commit only adds miniLZO to the build system of KernelStrider.

The subsequent patches will add LZO compression of the trace data in the
kernel, before the data are passed to the user space. They will be stored
in a compressed form there. Trace processors (tsan_process_trace, ...) will
then decompress the appropriate portions of the data and process them.

LZO was chosen for 2 reasons:
- LZO compression is already supported and used in the kernel;
- LZO compression is known to be fast, usually faster than other commonly
used compression techniques.