util: Fixed handling of events that wrap around the end of the buffer
Also, made sure the read and write positions do not grow indefinitely, just in
case.

Besides that, the test for this sitiation added before no longer writes the
text representation of the trace to a file but rather to /dev/null. It is not
needed anyway, the test only checks if all the compressed events have been saved
OK.