func: Disabled the test for kmalloc_order_trace
The test for kmalloc_order_trace is disabled because it is unclear how to
call this function reliably. It seems, it is not guaranteed to work for
memory blocks of <= KMALLOC_MAX_CACHE_SIZE but that may be subject to
change.

On OpenSUSE Tumbleweed with kernel 3.13.2 and SLAB allocator,
kmalloc_order_trace() returned such a value that resulted in a system crash
when kfree() was called for it.

On the other hand, when calling kmalloc_order_trace() with a larger value on
that system, it returned NULL and issued a warning to the system log (to be
exact, alloc_pages_* called from it eventually issued that).

SLUB-based system seem to be fine, although I can't be absolutely sure.

Anyway, KernelStrider will continue to intercept and process the calls to
kmalloc_order_trace(), it is just the test that has been disabled.