Disable CTF-based output subsystem by default
It needs a maintainer and seems to be not widely used anyway.
For it to be usable, some work is yet to be done.
For example, it should be updated to support kernel 3.15.

Not sure if it is worth investing in it further, disabling it for now.

One can set -DCTF_OUTPUT_SUBSYSTEM=on option to cmake to enable it.