core: Make code of the target writable for the time of instrumentation
Starting from upstream commit 4982223e51e8ea9d09bb33c8323b5ec1877b2b51 which
went into kernel 3.16, the code of a kernel module becomes read only before
the notification about MODULE_STATE_COMING triggers.

To be able to instrument the module anyway, we use the approach Ftrace
relies upon: temporarily make the code RW and make in RO again after the
instrumentation.

Currently, this is done for the kernel 3.16 and newer. If that upstream commit
is backported to older kernel (along with the appropriate changes in Ftrace),
some tests for KernelStrider may result in kernel oops. We'll need to fix
it then.