core: More fixes to make the system work with kernel 3.16
All changes to make KernelStrider work on 3.16 are needed because of the
following kernel commit:

  commit 4982223e51e8ea9d09bb33c8323b5ec1877b2b51
  Author: Rusty Russell <rusty@rustcorp.com.au>
  module: set nx before marking module MODULE_STATE_COMING.

Among other things, it changed the order of events: the notifications for
MODULE_STATE_COMING are processed earlier now than the information about the
target's ELF sections is sent to sysfs. Fortunately, it seems, KernelStrider no
longer requires this information, so I turned off the appropriate component.

The user mode helpers are no longer needed too.

Additionally, here is a fix for my commit d422fcc3c2e9. It is not only the code
of the target that becomes read-only, rodata becomes read-only too. As the jump
tables are commonly stored in rodata areas, it is needed to make them writable
too for the time of instrumentation. This is also done now.