Fix handling of the functions like __put_user_<N>
There are exported functions in the kernel that actually clobber
%ebx/%rbx despite any conventions.
 __put_user_<N> and __get_user_<N> are of this kind (see
arch/x86/lib/putuser.S from the kernel 4.1, for example).
They actually get away with that because are called from the inline
assembly and the compiler handles that clobber correctly.

No such luck for KernelStrider though, the modules under analysis crashed
because of that. Fortunately, these special functions do not clobber
other non-scratch registers. So we now use %edi on 32-bit systems and
%r15 on x86_64 to save %ebx/%rbx in kedr_thunk_* and get around the
clobber.