core: Try to instrument possibly incomplete functions as well
One special case when is_incomplete_function() may return non-zero is
when .exit.text is immediately followed by .altinstr_replacement
section.
The latter contains no functions, so the size of the last function in
.exit_text will be determined incorrectly, as if that function
included .altinstr_replacement as well. Unfortunately, we cannot
detect this case reliably yet.

The exit functions of cfg80211 and mac80211 modules in ROSA R6 x64 are
the examples of such conditions. Let us consider the code after the
actual end of the function "dead" and will try to instrument the
function anyway.